Fix GL_INVALID_OPERATION in font renderer when font cache is limited.
b/17332126

This patch is to fix a false logic in FontRenderer when
flashAllAndInValidate() happens. The flashAllAndInvalidate
will trigger an issueDrawCommand which will set mDrawn to
true even though there is no texture drawn. So that the
step of unbinding buffer in GL is skipped. And when the
second time issueDrawCommand get called, the
glVertexAttribPointer will be trigger with GL_ARRAY_BUFFER
wrongly bound and then finally lead to a GL_INVALID_OPERATION when
glDrawElements() in FontRenderer:issueDrawCommand get
called the second time.

Change-Id: I1a996354a84db911ace0e7b39ca77cd0dd3d1555

Signed-off-by: Raj Mamadgi <r.mamadgi@samsung.com>